refactor the current remainder of xsf functions into dedicated modules
