+++ src/alg.rs
@@ -1,0 +1,6 @@
+use crate::bindings;
+
+/// Cube root
+pub fn cbrt(x: f64) -> f64 {
+    unsafe { bindings::cbrt_(x) }
+}
