+++ src/binom.rs
@@ -1,0 +1,3 @@
+use crate::bindings::xsf_impl;
+
+xsf_impl!(binom, (n: f64, k: f64), "Binomial coefficient");
