+++ src/cdflib.rs
@@ -1,0 +1,3 @@
+use crate::bindings::xsf_impl;
+
+xsf_impl!(gdtrib, (a: f64, p: f64, x: f64), "Inverse of `p = gdtr(a, b, x)` with respect to `b`");
